refactor charts location in folder structure
